+++ src/SqlArtisan/Internal/SqlPart/Clause/ForUpdate/OfClause.cs
@@ -1,8 +1,13 @@
 namespace SqlArtisan.Internal;
 
-public sealed class OfClause(DbColumn tableIdentifier) : SqlPart
+public sealed class OfClause : SqlPart
 {
-    private readonly DbColumn _tableIdentifier = tableIdentifier;
+    private readonly DbColumn _tableIdentifier;
+
+    internal OfClause(DbColumn tableIdentifier)
+    {
+        _tableIdentifier = tableIdentifier;
+    }
 
     internal override void Format(SqlBuildingBuffer buffer) => buffer
         .Append(Keywords.Of)