+++ src/SqlArtisan/Internal/SqlPart/Condition/ArrayCondition/ArrayContainsCondition.cs
@@ -1,4 +1,7 @@
 namespace SqlArtisan.Internal;
 
-public sealed class ArrayContainsCondition(SqlExpression leftSide, SqlExpression rightSide) :
-    ArrayCondition(leftSide, Operators.ArrayContains, rightSide);
+public sealed class ArrayContainsCondition : ArrayCondition
+{
+    internal ArrayContainsCondition(SqlExpression leftSide, SqlExpression rightSide)
+        : base(leftSide, Operators.ArrayContains, rightSide) { }
+}