+++ src/SqlArtisan/Internal/SqlPart/Condition/JsonCondition/JsonbExistsAnyCondition.cs
@@ -1,4 +1,7 @@
 namespace SqlArtisan.Internal;
 
-public sealed class JsonbExistsAnyCondition(SqlExpression leftSide, SqlExpression[] keys) :
-    JsonbCondition(leftSide, Operators.JsonbExistsAny, new ArrayConstructorExpression(keys));
+public sealed class JsonbExistsAnyCondition : JsonbCondition
+{
+    internal JsonbExistsAnyCondition(SqlExpression leftSide, SqlExpression[] keys)
+        : base(leftSide, Operators.JsonbExistsAny, new ArrayConstructorExpression(keys)) { }
+}