+++ src/SqlArtisan/Internal/SqlPart/Expression/JsonOperator/JsonArrowOperator.cs
@@ -1,4 +1,7 @@
 namespace SqlArtisan.Internal;
 
-public sealed class JsonArrowOperator(SqlExpression leftSide, SqlExpression rightSide) :
-    BinaryOperator(leftSide, Operators.JsonArrow, rightSide);
+public sealed class JsonArrowOperator : BinaryOperator
+{
+    internal JsonArrowOperator(SqlExpression leftSide, SqlExpression rightSide)
+        : base(leftSide, Operators.JsonArrow, rightSide) { }
+}