+++ src/SqlArtisan/Internal/SqlPart/Expression/VectorOperator/L2DistanceOperator.cs
@@ -1,4 +1,7 @@
 namespace SqlArtisan.Internal;
 
-public sealed class L2DistanceOperator(SqlExpression leftSide, SqlExpression rightSide) :
-    BinaryOperator(leftSide, Operators.L2Distance, rightSide);
+public sealed class L2DistanceOperator : BinaryOperator
+{
+    internal L2DistanceOperator(SqlExpression leftSide, SqlExpression rightSide)
+        : base(leftSide, Operators.L2Distance, rightSide) { }
+}